Removed CSharpAdapter and IRunner
