+++ Game/scenes/worlds/main_test/main_test.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=14 format=3 uid="uid://fmn86ai2rkkq"]
 
-[ext_resource type="Script" path="res://src/Worlds/CSharpAdapter.cs" id="2_s2j2m"]
+[ext_resource type="Script" path="res://src/Worlds/MainTest/MainTestRunner.cs" id="1_t26kc"]
 [ext_resource type="PlaneMesh" uid="uid://sk1pvvyu4jlr" path="res://scenes/worlds/main_test/ground.tres" id="5"]
 [ext_resource type="Script" path="res://src/GUI/FreeCamera.gd" id="7"]
 
@@ -52,12 +52,10 @@ shadow_enabled = true
 transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, -32.3125, 6.14257, 16.0126)
 mesh = SubResource("4")
 
-[node name="CSharpAdapter" type="Node3D" parent="."]
-script = ExtResource("2_s2j2m")
-CreateFPSCounterUI = true
-CreateDebugInspector = true
+[node name="MainTestRunner" type="Node3D" parent="."]
+script = ExtResource("1_t26kc")
 
-[node name="Camera3D" type="Camera3D" parent="CSharpAdapter"]
+[node name="Camera3D" type="Camera3D" parent="MainTestRunner"]
 transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, -1.2747, 7.04908, 29.5765)
 script = ExtResource("7")
