--- Game/src/Worlds/IRunner.cs
@@ -1,9 +1,0 @@
-using Godot;
-
-namespace Worlds;
-
-public interface IRunner
-{
-    void Process(double delta);
-    void Input(InputEvent input);
-}
